fixed mapping

+++ app/src/main/java/com/example/ditest/domain/entity/Currency.kt
@@ -4,8 +4,8 @@ data class Currency(
     var iD: String,
     var numCode: String,
     var charCode: String,
-    var nominal: Long,
+    var nominal: Int,
     var name: String,
-    var value: Long,
-    var previous: Long
+    var value: Double,
+    var previous: Double
 )
